remove callback middleware, refactor unit tests
