+++ examples/serverless-vanilla/src/uppercase.js
@@ -1,0 +1,4 @@
+
+module.exports = () => ({
+  upperCase: input => input.toUpperCase()
+});
