+++ examples/serverless-vanilla/unit-test/app.test.js
@@ -1,9 +1,12 @@
-const handler = require("../handler").handler;
+const { FiveOhThreeError, ValidationError } = require("../src/errors");
+const upperCaseInstance = require("../src/uppercase");
+const app = require("../src/app");
 const comparisonHandler = require("../comparison_handler").handler;
 
 describe("handler", () => {
   let event;
   let context;
+  let upperCase;
 
   beforeEach(() => {
     // Minimal API Gateway Event via Lambda-Proxy
@@ -14,10 +17,18 @@ describe("handler", () => {
     // something down the road. At the very least, it
     // is an empty object IRL.
     context = {};
+    upperCase = jest.fn().mockResolvedValue();
   });
 
   it("should capitalize the `value` query param", async () => {
-    event.queryStringParameters.value = " SoMeThInG ";
+    const value = " SoMeThInG ";
+
+    // With LaconiaJS
+    await expect(app({ value }, upperCaseInstance())).resolves.toEqual({
+      value:  " SOMETHING "
+    });
+
+    event.queryStringParameters.value = value;
 
     const expectResponse = {
       body: '{"value":" SOMETHING "}',
@@ -26,9 +37,6 @@ describe("handler", () => {
       isBase64Encoded: false
     };
 
-    // With LaconiaJS
-    await expect(handler(event, context)).resolves.toEqual(expectResponse);
-
     // Without LaconiaJS
     await expect(comparisonHandler(event, context)).resolves.toEqual(
       expectResponse
@@ -36,15 +44,17 @@ describe("handler", () => {
   });
 
   it("should handle a 400 validation error", async () => {
+    // With LaconiaJS
+    await expect(app({}, upperCaseInstance())).rejects.toBeInstanceOf(
+      ValidationError
+    );
+
     const expectResponse = {
       body: "?value= is missing or empty.",
       headers: { "Content-Type": "text/plain" },
       isBase64Encoded: false,
       statusCode: 400
     };
-    // With LaconiaJS
-    await expect(handler(event, context)).resolves.toEqual(expectResponse);
-
     // Without LaconiaJS
     await expect(comparisonHandler(event, context)).resolves.toEqual(
       expectResponse
@@ -52,7 +62,14 @@ describe("handler", () => {
   });
 
   it("should handle a 503 unavailable error", async () => {
-    event.queryStringParameters.value = "503error";
+    const value = "503error";
+
+    // With LaconiaJS
+    await expect(app({ value }, upperCaseInstance())).rejects.toBeInstanceOf(
+      FiveOhThreeError
+    );
+
+    event.queryStringParameters.value = value;
 
     const expectResponse = {
       body: "Unsupported.",
@@ -63,9 +80,6 @@ describe("handler", () => {
       isBase64Encoded: false,
       statusCode: 503
     };
-
-    // With LaconiaJS
-    await expect(handler(event, context)).resolves.toEqual(expectResponse);
 
     // Without LaconiaJS
     await expect(comparisonHandler(event, context)).resolves.toEqual(
